1 - Added boss contact damage scaling with boss velocity. 2 - Nightwither added to Calamity Dust enum. 3 - Providence night variant and spawn animation added. 4 - A lot of new grey textures for the Prov night variant. 5 - Removed some useless casts in various files. 6 - Profaned Cores can now be used during nighttime and changed the tooltips to be ambiguous. 7 - Bobbit Worm lunge speed buffed. 8 - Buffed several abyss enemy HP and damage. 9 - Fixed Cuttlefish and some other abyss enemies not properly utilizing abyss stealth mechanics. 10 - Fixed several flying worm enemies that were meant to be passive until attacked that weren't passive at all, most notably the Sea Serpent. 11 - Removed some unnecessary code in several worm NPC files. 12 - Mirage Jellies should teleport more sensibly now. 13 - Edited Providence's AI a lot to account for the new nighttime AI and spawn animation. 14 - Yharon AI completely reworked, mostly phase 1. HAS NOT BEEN TESTED BY ME YET. 15 - Brimstone Monsters spread out speed cut in half. 16 - Rev+ Plantera spawns an extra set of tentacles when below 10% HP, however, her defense only scales up to 99 in this phase, not 9999. 17 - Death Mode ocean sharknadoes and cthulhunadoes no longer spawn sharkrons. 18 - Added new method for Tenebris to spread and Lumenyl/Sea Prism Crystals to spawn, in CalamityWorld. 19 - Added GetLerpValue method from Terraria 1.4 to CalamityUtils, can be removed when necessary.

+++ NPCs/Polterghast/PhantomFuckYou.cs
@@ -74,7 +74,7 @@ namespace CalamityMod.NPCs.Polterghast
             }
 
 			// Percent life remaining, Polter
-			float lifeRatio = (float)Main.npc[CalamityGlobalNPC.ghostBoss].life / (float)Main.npc[CalamityGlobalNPC.ghostBoss].lifeMax;
+			float lifeRatio = Main.npc[CalamityGlobalNPC.ghostBoss].life / Main.npc[CalamityGlobalNPC.ghostBoss].lifeMax;
 
 			npc.TargetClosest(true);
 
@@ -96,7 +96,7 @@ namespace CalamityMod.NPCs.Polterghast
             }
 
             Player player = Main.player[npc.target];
-            double deg = (double)npc.ai[1];
+            double deg = npc.ai[1];
             double rad = deg * (Math.PI / 180);
             double dist = 500;
             npc.position.X = player.Center.X - (int)(Math.Cos(rad) * dist) - npc.width / 2;
@@ -125,7 +125,7 @@ namespace CalamityMod.NPCs.Polterghast
 				spriteEffects = SpriteEffects.FlipHorizontally;
 
 			Texture2D texture2D15 = Main.npcTexture[npc.type];
-			Vector2 vector11 = new Vector2((float)(Main.npcTexture[npc.type].Width / 2), (float)(Main.npcTexture[npc.type].Height / 2));
+			Vector2 vector11 = new Vector2(Main.npcTexture[npc.type].Width / 2, Main.npcTexture[npc.type].Height / 2);
 			Color color36 = Color.White;
 			float amount9 = 0.5f;
 			int num153 = 5;
@@ -137,16 +137,16 @@ namespace CalamityMod.NPCs.Polterghast
 					Color color38 = lightColor;
 					color38 = Color.Lerp(color38, color36, amount9);
 					color38 = npc.GetAlpha(color38);
-					color38 *= (float)(num153 - num155) / 15f;
-					Vector2 vector41 = npc.oldPos[num155] + new Vector2((float)npc.width, (float)npc.height) / 2f - Main.screenPosition;
-					vector41 -= new Vector2((float)texture2D15.Width, (float)(texture2D15.Height)) * npc.scale / 2f;
+					color38 *= (num153 - num155) / 15f;
+					Vector2 vector41 = npc.oldPos[num155] + new Vector2(npc.width, npc.height) / 2f - Main.screenPosition;
+					vector41 -= new Vector2(texture2D15.Width, texture2D15.Height) * npc.scale / 2f;
 					vector41 += vector11 * npc.scale + new Vector2(0f, 4f + npc.gfxOffY);
 					spriteBatch.Draw(texture2D15, vector41, npc.frame, color38, npc.rotation, vector11, npc.scale, spriteEffects, 0f);
 				}
 			}
 
 			Vector2 vector43 = npc.Center - Main.screenPosition;
-			vector43 -= new Vector2((float)texture2D15.Width, (float)(texture2D15.Height)) * npc.scale / 2f;
+			vector43 -= new Vector2(texture2D15.Width, texture2D15.Height) * npc.scale / 2f;
 			vector43 += vector11 * npc.scale + new Vector2(0f, 4f + npc.gfxOffY);
 			spriteBatch.Draw(texture2D15, vector43, npc.frame, npc.GetAlpha(lightColor), npc.rotation, vector11, npc.scale, spriteEffects, 0f);
